edit: fixed bug's and finilized the analytic's endpoint's in the transaction's module

+++ finance_backend/app/transactions/schemas.py
@@ -90,5 +90,5 @@ class AccountBalance(BaseModel):
     balance: Decimal
 
 class AccountBalancesOut(BaseModel):
-    total_balance: int
+    total_balance: Decimal
     accounts: List[AccountBalance]
